Move engine out of SheetEditor.
